requirements.txt
wheel==0.46.3

absl-py==2.3.1
addict==2.4.0
ai-edge-litert-nightly==2.2.0.dev20260210
ai-edge-quantizer-nightly==0.5.0.dev20260212
ai-edge-tensorflow==2.21.0.dev20251110
aiofiles==24.1.0
alembic==1.17.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appdirs==1.4.4
asttokens==3.0.1
astunparse==1.6.3
backports-strenum==1.3.1
black==25.12.0
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.2
certifi==2025.11.12
cityscapesscripts==2.2.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.2
crcmod==1.7
cryptography==46.0.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.5
cycler==0.12.1
databricks-sdk==0.73.0
debugpy==1.8.20
decorator==5.2.1
decord==0.6.0
docker==7.1.0
einops==0.8.1
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.124.4
ffmpy==1.0.0
filelock==3.14.0
fire==0.7.1
flask-cors==6.0.1
flask==3.1.2
flatbuffers==25.12.19
fonttools==4.60.1
fsspec==2025.9.0
ftfy==6.1.1
fvcore==0.1.5.post20221221
gast==0.7.0
gitpython==3.1.45
google-auth==2.43.0
google-pasta==0.2.0
gradio-client==2.0.2
gradio==6.2.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.2.4
groovy==0.1.2
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huey==2.5.4
huggingface-hub==0.36.0
humanfriendly==10.0
hydra-core==1.3.2
imageio==2.37.2
immutabledict==4.3.0
importlib-metadata==8.7.0
iniconfig==2.3.0
iopath==0.1.10
ipykernel==7.2.0
ipython==8.38.0
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
jaxtyping==0.3.7
jedi==0.19.2
jmespath==0.10.0
joblib==1.5.2
jupyter-client==8.8.0
jupyter-core==5.9.1
kagglehub==0.4.3
kagglesdk==0.1.15
keras-nightly==3.12.0.dev2025100703
kiwisolver==1.4.9
libclang==18.1.1
mako==1.3.10
mamba-ssm==2.3.0
markdown-it-py==4.0.0
markdown==3.10
markupsafe==2.1.5
matplotlib-inline==0.2.1
matplotlib==3.10.7
mdurl==0.1.2
ml-dtypes==0.5.4
mlflow-skinny==3.6.0
mlflow-tracing==3.6.0
mlflow==3.6.0
model-index==0.1.11
multipledispatch==1.0.0
mypy-extensions==1.1.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.13.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
nvitop==1.6.1
omegaconf==2.3.0
open-clip-torch==3.2.0
opencv-python==4.12.0.88
opendatalab==0.0.10
openmim==0.3.9
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
openxlab==0.1.3
opt-einsum==3.4.0
optree==0.18.0
ordered-set==4.1.0
orjson==3.11.5
oss2==2.17.0
packaging==24.2
pandas==2.3.3
parso==0.8.6
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
pluggy==1.6.0
portalocker==3.2.0
prettytable==3.17.0
prompt-toolkit==3.0.52
protobuf==6.33.1
psutil==7.1.3
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==22.0.0
pyasn1-modules==0.4.2
pyasn1==0.6.1
pycocotools==2.0.10
pycryptodome==3.23.0
pydub==0.25.1
pygments==2.19.2
pyparsing==3.2.5
pyquaternion==0.9.9
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytokens==0.3.0
pytz==2023.4
pyzmq==27.1.0
regex==2025.11.3
requests==2.28.2
rich==13.4.2
rsa==4.9.1
safehttpx==0.1.7
safetensors==0.7.0
sam2==1.1.0
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.1
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.44
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.50.0
tabulate==0.9.0
tb-nightly==2.20.0a20250717
tensorboard-data-server==0.7.2
tensorboard==2.20.0
termcolor==3.2.0
tf-nightly==2.22.0.dev20260211
threadpoolctl==3.6.0
timm==1.0.22
tokenizers==0.22.1
tomli==2.3.0
tomlkit==0.13.3
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
triton==3.6.0
typer-slim==0.20.0
typer==0.20.1
typing==3.7.4.3
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.38.0
uvloop==0.22.1
wadler-lindig==0.1.7
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
werkzeug==3.1.3
wrapt==2.1.1
yacs==0.1.8
yapf==0.43.0